Don't run eholdings test suite on deploy.
Whenever this build is run, we lint the eholdings source, and then run
the eholdings suite tests before deploying. This takes several minutes
as can be seen here:

https://travis-ci.org/folio-org/ui-eholdings/jobs/330103495

But this linting and testing work isn't necessary however because the
test suite is being run as part of the actual repo build. In fact,
because master is protected by the test suite on the eholding
repository, it's impossible for any changes to actually be seen by
this deployment repo unless they've already run (and passed) the test
suite.

This removes those checks as redundant so that we can have faster
merge-to-deployment times.